chore: refactor existing DownloadButton to separate component from logic
